requirements.txt
Flask
pymongo
python-dotenv
dnspython
bson